+++ APBDFinalProject/APBDFinalProject/Models/RequestModels/BusinessCustomerRequest.cs
@@ -2,7 +2,7 @@ namespace APBDFinalProject.RequestModels;
 
 public class BusinessCustomerRequest
 {
-    public int KRS { get; set; }
+    public long KRS { get; set; }
     public string BusinessName { get; set; }
     public string Address { get; set; }
     public string Email { get; set; }
